jsonrpc: fix client response stream queue pop bug
This fixes a bug that happened on the client side of a stream, when
multiple messages were received and queued for processing in a short
time. The queue elements were not properly popped, causing the first
element to be processed multiple times.

A test is added to lock in the correct behavior.