clientrpc: add replaymsglog-related interface
This adds the necessary messages and methods to control the replaymsglog
feature for PMs and GCMs in the client's RPC server.

In particular, it allows fetching unacked messages from the server
when initializing a stream and also to ack processed messages to remove
them from the replaymsg store.

The client rpcserver implementation is modified with dummy
implementation for the methods.