client: automatically reset KXs if offline for too long
This adds a check after connecting to the server to verify if the local
client has been offline for longer than the server's message retention
policy. If true, then the client sends a reset KX command to all known
users.

This is needed to prevent a situation where, after being offline for too
long, all of the client's ratchets are broken due to lost messages.