multi: return generic error on C2S RM payment failures
This makes the RouteMessage command fail with a generic payment error
on all payment-related failures. This makes it more obvious to the
client that the issue is related to the payment and that it should
attempt to send the payload again after fetching and paying for a new
invoice.

This fixes a possible situation where the message is pushed but not
acked and the client tries to continuously reuse the payment, leading to
connection loops.