multi: make post sub/unsub async
This makes the calls to Subscribe/Unsubscribe to posts async, instead of
sync. This improves the case where the remote user is offline and will
take long to reply to the subcription request.

New events are added to notify the UI about a change in subscrption
status.

Both brclient and bruig are updated to handle the new subscription
system.

Fix #54 #68.