server: support multiple concurrent RM pushes
This adds support in the server to multiple concurrent RM pushes.
Instead of a single invoice that needs to be paid by clients, the server
accepts the ID of a previously paid invoice, up to a configurable
maximum of outstanding invoices (per connection).

This allows parallel sending of RMs by compatible clients.

Clients that do not support parallel sending (i.e. all clients up to a
future commit that will implement parallel sending) can still connect
and send messages to the server, as a special compatibility test is
performed when no invoice id is provided with a push.