client: track and warn unsupported GC version
This adds a new Version field in GC definitions and warns the user when
a GC list with an unsupported version is received. This will be used in
the future to warn about upgraded GCs.

The current GC version is assumed to be version 0.

A function is added in the clients to resend the GC list, to fix issues
with the list getting out of sync in members.