jsonrpc: improve peer.readLoop() shutdown
This adds an inner context in readLoop which gets canceled in case of
unexpected errors and lets requests being processed finish their
processing earlier.

EOFs are handled without canceling the inner context to account for
POST-based peers, where only one (or a few) commands are read and then
followed by EOF.

A maximum of one second is allowed for any outstanding requests to
finish after an error is detected in readLoop.