client: persist and retry TipUser
This adds persistence and retries to the TipUser() client action. It
makes the action asynchronous, therefore the signature of the function
was changed to reflect that.

Callers can specify the max amount of retries to attempt when sending a
tip. The existing clients are set to use a max attempt of 1, while
clientrpc users may specify the max attempts with a new field.