client: fix concurrent autokx causing broken ratchets
This fixes a bug when two users attempt to autokx concurrently, causing
a broken ratchet due to both ratchets completing but each user selecting
a different one to use.

One situation this can happen is when the users are added to two
different GCs, such that they are already on the other one, thus both
attempt to KX with the other simultaneously.

This was also the cause for some recent CI errors.

The fix is to prevent one of the users from continuing one of the KX
attempts, thereby ensuring only one KX completes and its ratchet
selected for using.

The selection of which ratchet to use is done at random, by selecting
to stop the one that has a smaller initial RV number.

An E2E test is added to simulate the issue and verify the fix works.