client: add new onboard stage to track outbound ln mine height
The new stage detects when the outbound channel was mined and tracks
that height. The existing stage that tracks the confirmation of the
outbound channel now sends an update after every block, showing how many
confirmations are left before the channel is usable.

brclient and bruig are updated to show the new information.