client: fix buggy autounsub from older clients
This fixes a bug that could make older clients trigger the auto
unsubscription of idle users in a loop.

The issue could be triggered in clients that did not have commit
15690ddfac057bd2ece38b110ba559d7277c2663 applied (i.e. clients older
than version v0.2.0).

After that commit, the last handshake attempt time is cleared from the
DB after the handshake completes. This is later taken into account in
the autounsub logic, to detect when to attempt a new handshake.

However, clients that were online and performed handshakes before that
commit would still record the old handshake attempt time, even for
successful handshakes. Later on, the autounsub logic would fail to
attempt a _new_ handshake (assuming the prior one failed).

This commit fixes the issue by detecting the faulty state: a decryption
time that is more recent than the last attempted handshake time
(implying the local client received a SYN or ACK message). In this
situation, it can be assumed that the handshake was completed
successfully.

An integration test is added to assess the fix.