client: handle PMs and GCMs in RV manager goroutine
This ensures that received PMs and GCMs are processed sequentially. In
particular, each remote user's PMs and GCMs will be processed in the
sequence that they were sent, as determined by the sequential nature of
consecutive RV points for a ratchet.

This should ensure that the GCM cacher receives messages in the order
they were sent in situations where many messages are coming in (for
example, after restarting the client after a long period offline where
messages are coming in).