client: support bundled resources
A bundled resource compiles several resources together into a single
reply to a resource. When the remote client attempts to fetch one of the
bundled resources, then that resource is fetched locally instead of
performing a remote request.

This is skipped if the request has data (i.e. it's a POST). The bundles
are only checked on a per-session fashion (new sessions cause a new
bundle download).